+++ pmd-core/src/main/java/net/sourceforge/pmd/Report.java
@@ -19,6 +19,7 @@ import java.util.Set;
 
 import org.apache.commons.lang3.StringUtils;
 
+import net.sourceforge.pmd.annotation.Experimental;
 import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.dfa.report.ReportTree;
 import net.sourceforge.pmd.lang.rule.stat.StatisticalRule;
@@ -26,6 +27,7 @@ import net.sourceforge.pmd.renderers.AbstractAccumulatingRenderer;
 import net.sourceforge.pmd.stat.Metric;
 import net.sourceforge.pmd.util.DateTimeUtil;
 import net.sourceforge.pmd.util.NumericConstants;
+import net.sourceforge.pmd.util.Predicate;
 
 /**
  * A {@link Report} is the output of a PMD execution. This
@@ -34,6 +36,15 @@ import net.sourceforge.pmd.util.NumericConstants;
  * a report (see {@link PmdAnalysis#performAnalysisAndCollectReport()})
  * The mutation methods on this class are deprecated, as they will be
  * internalized in PMD 7.
+ *
+ * <p>For special use cases, like filtering the report after PMD analysis and
+ * before rendering the report, some transformation operations are provided:
+ * <ul>
+ *     <li>{@link #filterViolations(Predicate)}</li>
+ *     <li>{@link #union(Report)}</li>
+ * </ul>
+ * These methods create a new {@link Report} rather than modifying their receiver.
+ * </p>
  */
 public class Report implements Iterable<RuleViolation> {
 
@@ -681,5 +692,70 @@ public class Report implements Iterable<RuleViolation> {
     @Deprecated
     public void addListeners(List<ThreadSafeReportListener> allListeners) {
         listeners.addAll(allListeners);
+    }
+
+    /**
+     * Creates a new report taking all the information from this report,
+     * but filtering the violations.
+     *
+     * @param filter when true, the violation will be kept.
+     * @return copy of this report
+     */
+    @Experimental
+    public Report filterViolations(Predicate<RuleViolation> filter) {
+        Report copy = new Report();
+        copy.start = start;
+        copy.end = end;
+
+        for (RuleViolation violation : violations) {
+            if (filter.test(violation)) {
+                copy.addRuleViolation(violation);
+            }
+        }
+
+        copy.linesToSuppress.putAll(linesToSuppress);
+        copy.suppressedRuleViolations.addAll(suppressedRuleViolations);
+        copy.metrics.addAll(metrics);
+        copy.errors.addAll(errors);
+        copy.configErrors.addAll(configErrors);
+        return copy;
+    }
+
+    /**
+     * Creates a new report by combining this report with another report.
+     * This is similar to {@link #merge(Report)}, but instead a new report
+     * is created. The lowest start time and greatest end time are kept in the copy.
+     *
+     * @param other the other report to combine
+     * @return
+     */
+    @Experimental
+    public Report union(Report other) {
+        Report copy = new Report();
+
+        copy.start = Math.min(other.start, this.start);
+        copy.end = Math.max(other.end, this.end);
+
+        for (RuleViolation violation : violations) {
+            copy.addRuleViolation(violation);
+        }
+        for (RuleViolation violation : other.violations) {
+            copy.addRuleViolation(violation);
+        }
+
+        copy.linesToSuppress.putAll(linesToSuppress);
+        copy.linesToSuppress.putAll(other.linesToSuppress);
+
+        copy.suppressedRuleViolations.addAll(suppressedRuleViolations);
+        copy.suppressedRuleViolations.addAll(other.suppressedRuleViolations);
+
+        copy.metrics.addAll(metrics);
+        copy.metrics.addAll(other.metrics);
+        copy.errors.addAll(errors);
+        copy.errors.addAll(other.errors);
+        copy.configErrors.addAll(configErrors);
+        copy.configErrors.addAll(other.configErrors);
+
+        return copy;
     }
 }