+++ docs/pages/release_notes.md
@@ -23,6 +23,7 @@ This is a {{ site.pmd.release_type }} release.
 
 ### Fixed Issues
 * core
+  * [#3792](https://github.com/pmd/pmd/issues/3792): \[core] Allow to filter violations in Report
   * [#3881](https://github.com/pmd/pmd/issues/3881): \[core] SARIF renderer depends on platform default encoding
   * [#3882](https://github.com/pmd/pmd/pull/3882): \[core] Fix AssertionError about exhaustive switch
   * [#3884](https://github.com/pmd/pmd/issues/3884): \[core] XML report via ant task contains XML header twice
@@ -37,6 +38,14 @@ This is a {{ site.pmd.release_type }} release.
   * [#3706](https://github.com/pmd/pmd/issues/3706): \[plsql] Parsing exception CURSOR statement with parenthesis groupings
 
 ### API Changes
+
+#### Experimental APIs
+
+* Report has two new methods which allow limited mutations of a given report:
+  * {% jdoc !!core::Report#filterViolations(net.sourceforge.pmd.util.Predicate) %} creates a new report with
+    some violations removed with a given predicate based filter.
+  * {% jdoc !!core::Report#union(net.sourceforge.pmd.Report) %} can combine two reports into a single new Report.
+* {% jdoc !!core::util.Predicate %} will be replaced in PMD7 with the standard Predicate interface from java8.
 
 ### External Contributions
 * [#3883](https://github.com/pmd/pmd/pull/3883): \[doc] Improve side bar by Adding Release Date - [@jasonqiu98](https://github.com/jasonqiu98)